renamed .ts files to .mts and made changes to make things work
