+++ config/eslint/eslintConfig.json
@@ -16,19 +16,22 @@
 
   "overrides": [
     {
-      "files": ["src/RollbarClient*.ts"],
+      "files": ["src/RollbarClient*.mts"],
       "env": {
         "browser": true
       },
       "rules": {
         "@typescript-eslint/unbound-method": "off",
+        "import/extensions": "off",
         "no-console": ["error", { "allow": ["debug", "error", "info", "warn"] }],
         "unicorn/filename-case": ["error", { "case": "pascalCase" }]
       }
     },
     {
-      "files": ["./src/**/*.test.ts"],
+      "files": ["./src/**/*.test.mts"],
       "rules": {
+        "import/no-extraneous-dependencies": "off",
+        "import/no-unresolved": "off",
         "unicorn/filename-case": "off"
       }
     }
