+++ src/RollbarClientSubmitter.test.mts
@@ -4,8 +4,12 @@ import * as ErrorStackParser from 'error-stack-parser';
 
 // Internal Imports
 import { name as expectedlibraryName, version as expectedlibraryVersion } from '../package.json';
-import { RollbarClientSubmitter } from './RollbarClientSubmitter';
-import type { IConfigurationOptions, IGenericObjectIndexSignature, IPayload } from './interfaces';
+import { RollbarClientSubmitter } from './RollbarClientSubmitter.mjs';
+import type {
+  IConfigurationOptions,
+  IGenericObjectIndexSignature,
+  IPayload,
+} from './interfaces.mjs';
 
 // Local Variables
 const acceptedLogLevels = ['critical', 'debug', 'error', 'info', 'warning'];
