+++ src/interfaces.mts
